+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+
+## UNRELEASED
+
+### Added
+
+- DB (database/sql) and HTTP Health Checkers
+
+
 ## 0.1.0 - 2017-02-08
 
 - Initial release